Removed UI tests on CI - TOO PRICEY
UI tests take too long to finish on the CI due to it running an emulator. This uses up a lot of my free GitHub Actions minutes which I can't afford.

From now on, UI tests should be run locally before pushing to your branch or merging to main.

Moreover, with the compose migration and how I am hoisting up state to the VM there, the importance for UI tests decrease dramatically.
This is thanks to hoisting state to the VM which allows me to test more in the Unit Tests instead.